don't overwrite incoming test-harness/.env

+++ test-harness.sh
@@ -1,13 +1,13 @@
 #!/usr/bin/env bash
 
-START=$(date "+%s")
-
 set -e
+
+START=$(date "+%s")
+THIS=$(basename "$0")
+
 
 ENV_FILE=".test-env"
 source $ENV_FILE
-
-echo "test-harness.sh: AFTER sourcing from $ENV_FILE. Build TYPE=$TYPE"
 
 rootdir=$(dirname "$0")
 pushd "$rootdir"
@@ -19,25 +19,29 @@ if [ -d "$SDK_TESTING_HARNESS" ]; then
   popd
   rm -rf "$SDK_TESTING_HARNESS"
 else
-  echo "test-harness.sh: directory $SDK_TESTING_HARNESS does not exist - NOOP"
+  echo "$THIS: directory $SDK_TESTING_HARNESS does not exist - NOOP"
 fi
 
 git clone --depth 1 --single-branch --branch "$SDK_TESTING_BRANCH" "$SDK_TESTING_URL" "$SDK_TESTING_HARNESS"
 
-## OVERWRITE incoming .env with .test-env
-cp "$ENV_FILE" "$SDK_TESTING_HARNESS"/.env
+
+if [[ $OVERWRITE_TESTING_ENVIRONMENT == 1 ]]; then
+  ## OVERWRITE incoming .env with .test-env
+  echo "$THIS: OVERWRITE downloaded $SDK_TESTING_HARNESS/.env with $ENV_FILE:"
+  cp "$ENV_FILE" "$SDK_TESTING_HARNESS"/.env
+fi
 
 ## Copy feature files into the project resources
 rm -rf tests/features
 mkdir -p tests/features
 cp -r "$SDK_TESTING_HARNESS"/features/* tests/features
-echo "test-harness.sh: seconds it took to get to end of cloning + copying: " + $(($(date "+%s") - $START))
+echo "$THIS: seconds it took to get to end of cloning + copying: " + $(($(date "+%s") - $START))
 
 ## Start test harness environment
 pushd "$SDK_TESTING_HARNESS"
 ./scripts/up.sh
 popd
-echo "test-harness.sh: seconds it took to finish testing sdk's up.sh: " + $(($(date "+%s") - $START))
+echo "$THIS: seconds it took to finish testing sdk's up.sh: " + $(($(date "+%s") - $START))
 echo ""
 echo "--------------------------------------------------------------------------------"
 echo "|"
